Use the JFET to switch the signal, not pull it to ground
I really thought I was being clever by combining the voluem control with
pulling the signal to the ground through the transistor.  But it turns
out I just can't fix the noise that this introduces by basically making
the transistor voltage be part of the path to ground.

Maybe it's fixable some other way, but I have seen the error of my "try
to be clever" ways, and am instead just using the JFET in a much simpler
capacity as just a direct analog switch.  That seems to work much
better.

That does mean that now the "depth" knob that used to control the signal
modulation strength is just a passive volume knob, and is as such
entirely useless.

The "clever" part of the pull-down-to-ground thing was how I avoided any
signal blending circuitry, because it was a "sometimes we pull down to
ground, sometimes we don't" and the blend pot just changed how _hard_ it
was pulled down to ground.

Now I'll have to replace that with a "how much to blend in the original
signal" instead, but at least I know how to do that, even if this commit
doesn't do it yet.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>