Clean up and finish re-org of the the octaver pedal too
The octaver pedal is trying to no longer be so much of a fuzz, and
trying much harder to be a cleaner octave-up pedal instead.  And I've
simulated it more to try to make it so, and that we don't have any
nassty oscillation issues.

Whether those simulations pan out with better sounds, we will have to
just wait and see..

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>