Sometimes I have the mental capacity of a hamster on crack
I'm not making this up - I was looking at the reverse bias leakage of
the JFET, because you can see it show up when the gate voltage goes
sufficiently positive.  And in my infinite wisdom I mentally went "maybe
I should put a diode on the control voltage", and even ended up deciding
to see what that looked like in simulation.

Not my proudest moment.

Only when it made no difference did my reptilian brain go "D'oh".

Let's just change the LFO to oscillate between ground and -V rather than
the whole voltage swing, when we don't actually want to reverse bias
this poor JFET.

And then after that I simulated my old circuit - pre op-amp - and it
worked just fine.  That was a bit odd, becasue I didn't expect this to
really matter, so I went back and forth on the LFO voltage swing, and it
really doesn't matter all that much.

Sure, the negative-only LFO was better, but it wasn't enormously better,
We're talking tens of uV in simulation.  uV, not mV.  And either way the
"make JFET pull to ground" circuit was doing really well.

That was strange too, because the reason I had done the op-amp thing in
the first place was that I couldn't get to those kinds of numbers with
that circuit.

Or so I thought.

I don't know what my problem was.  I'm wondering whether I saw those uV
numbers, and didn't react to the 'u' part, and thought I was still in
the mV range.  Regardless, I've gone back and simulated things again,
and the old simpler JFET circuit simply looks better.

[ For next time this happens: I think I got confused because I simulated
  a 1mV signal and compared that to something a bit saner.

  Simulate no signal at all to see just the switching noise, and a ~0.1V
  amplitude signal for a "somewhat realistic guitar volume", but a 1mV
  signal makes little sense ]

I must have had some hamster-brain moment, but now I can't replicate my
problems from before and I'm not sure what I changed.  I was playing
around tweaking the LFO resistor values, and tweaking the signal for
simulation, and I've clearly lost the plot.

[ Twilight Zone theme music plays in the background

  Or maybe I _still_ have a hamster-brain moment and I'm missing
  something big and this is all just wrong. ]

The only solution is obviously to just have both versions manufactured,
and let reality fight it out, Mad Max style.

[ Twilight Zone theme music changes to Tina Turner singing 'We Don't
  Need Another Hero'. "Two circuits enter, one circuit leaves" ]

This whole circuit really hasn't been working all that well, but it sure
is a teaching moment.  Sometimes the moment is "That's how this
transistor works".  And sometimes it is more along the lines of "Kids,
don't do drugs.  This is your hamster brain on crack"

The definition of insanity is doing the same thing over and expecting
different results.  But here we are anyway.

The guitar pedal kits I did tended to have a vactrol in them, and I
really didn't want to do that, and still don't.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>