Add PT2399-based delay pedal design
This is pretty much the datasheet surround/delay design, just using the
VGND from the base board instead of the built-in resistor divider
circuit of the chip.

Similarly, this also doesn't re-bias the input/output lines, since they
are already biased to VREF.

Let's see if I intepreted the datasheet correctly.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>