Update commentary on the delay effect low-pass filters
I misread the low-pass filters on input and output - since they are in
the op-amp feedback loop, they aren't simple passive filters, but
multiple-feedback active ones (not too dissimilar to the one I ended up
using in the octaver circuit, in fact).

And that makes a big difference to the frequency calculations.  I didn't
simulate it, but the calculator at

   http://sim.okawa-denshi.jp/en/OPttool.php

says it's a unity gain inverting filter (as expected) with the cut-off
frequency at 7.15kHz and a Q of ~1.1.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>