Make a switchable combo version of the USB-PD power board
So this was triggered partly by me noticing that my USB PD sources are
very inconsistent when it comes to 12V output that the original board
requested.

Several of my wall chargers do 15V but not 12V.  Of course, then some of
my battery packs are the reverse, and do 12V but not 15V output.

So there is no single thing that works for all common cases.  But 15V is
the more common option, and is required by the PD specs for anything
that claims to be 45W or more.  12V is entirely optional.

End result: add a switch to pick whether the power board requests 12V or
15V from the USB-C PD source.

And since 15V is now an option, add another LDO, and a switch to make
that LDO generate either 9V or 12V output (as per commit eb97091: "Add a
USB-C pedal power board" that already mentioned this option).

Connect the LDOs to two ports each, which also results in a certain
amount of isolation.

End result: the first two ports always output 9V, while the second two
ports can optionally provide 12V output if your USB-PD source does 15V.

The "switch" is either just a pin header with a jumper, or one of the
small slide switches that I have.  We're not doing anything fancy here.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>