Move mono connectors out by 0.8mm each
Minimal trivial fixup to make the mono connectors poke out of the
enclosure just a tad more: this way we should have enough threads to
actually have the nut bite down on the jack.

Is the nut required? No.  But it should make things a bit stabler, and
let's do it right.

Is this enough? We'll see, but from a quick test-fit it looks just
barely sufficient, and I don't want to overdo it.

This is not worth doing new boards for, but I'm fixing it now so that
the next time I order more boards I will have this fixed.  I've meant to
do this for a long time, but just never got around to it.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>