Add "VIC-20" effects board
This is a silly board just for fun, because of the whole "digital" board
thing.

We all know that the pinnacle of computing was the Commodore VIC-20, and
it even had sound effects.  So let's do a guitar pedal in honor of that.

The way to do that, is to find the absolute crappiest simple cheap 8-bit
microcontroller that has an ADC and DAC.  Of course, I didn't spend very
much time searching, but the ATTiny814 certainly seems to fit the bill.

This is going to be really really bad.

Oh, and I still use the OPA1655 and OPA1656 in these circuits, but I
really need to go find a better op-amp for the 5V range.  The OPA165x
"works" in that range, but it's kind of borderline.  There should be
lots of cheaper op-amps with 5V limits that do proper rail-to-rail
inputs, because the OPA165x is RtR only on output.

Off to parametric search I go.  The good news is that at least the
footprints are all standard, so this doesn't affect board layout.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>