Update USB-C pedal power board a bit more
Testing the old 12V-only board shows that several of my PD power
supplies that claim to only do 5V, 9V and 15V actually do 12V even if
they don't say so on the power supply.

It looks like 12V is common mainly due to Qualcomm Quick Charge (where
it is part of the spec), and so if the PD power supply does that, it
likely does 12V USB-PD too.

But having the switchable option is good regardless, and this adds a
updates the capacitor choices a bit.

I don't really have a great way to test power quality, but waving the
output end of the guitar pedal power supply next to a guitar cable that
is connected to the amp at the other end seems to be a good first
sanity-check.

And according to that simple check, the first (12V-only) pedal board is
fine.  Better than a commercial one I have around, in fact.  So these
may be quite usable.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>