Fix up some mis-designs in the the digital base board
First off, the LED power really should come from the +9V rail, not from
the LDO.  On the original VGND board, this didn't matter, since VIN was
the +9V rail with just the reverse voltage protection etc, but on this
board with the LDO, it's just wasting an LDO to make it drive the LED.

And worse than wasting the LDO on it, it makes the LED resistor overly
sensitive to the actual LED voltage, be it some red LED with a voltage
drop of 1.7V or a blue LED that drops 3.2V.

With the 9V rail, that's not such a huge deal, but with a 5V rail it's a
noticeable part of the whole available voltage and makes a big
difference for the current limiting resistor.

Ok, not really a fatal thing, and the fact that a super-efficient red
LED typically wants much less current than some fancy - but much less
efficient - purple LED would need means that you still do want to pick
the right resistor value, and the one on the schematic is just a
starting point.  But it was still not a good choice.

And I really need to sit down and learn more about the proper rules of
capacitive loads of op-amps and how to do this all right, but I'm pretty
sure I got the resistor very wrong on the output path to the capacitor.
This circuit does *not* want to slow down the charging and discharging
of the output capacitor, that just messes with the signal.

Now, I'm not sure I actually need the resistor to VREF - I think the
op-amp will drive the right voltage, but I need to think about it more.
In the meantime, add the resistor to VREF to make this look more like
the case I know works.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>