Update some of the rectifier values and commentary
This board actually sounds fun, and I have positive feedback from an
actual real guitar player.  But the "Tone" knob is mislabeled, and the
range of it should be adjusted.

It's more properly called "Weird", and instead of the normal flat Q=1
being in the middle, that boring normal setting should be at the
beginning of the pot, and it should go weird from there.

In the old model, going the other way (turning the tone knob down from
the middle) really didn't do anything interesting, so half of the range
of the knob was pretty useless.

But other than that kind of fairly minor details, it's been a success in
this form.

I think the range of the Weird pot could/should probably still be played
with, and in particular I suspect it might be worth looking into adding
resistors between the wiper and the ends to allow modifying the taper of
the pot to make the interesting range more useful.

But that's for a future date.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>