Let's read the PT2399 datasheet, ok?
I'm sure I read the part where it said to use a 100uF capacitor, but I
clearly didn't internalize it.  And maybe it's overkill, but that thing
does claim to use up to 100mA, which is certainly more than any of my
other effects boards have ever approached.

Technically, I think we only really need one on the post-LDO +5V rail,
which is what the PT2399 obviously uses.  But I don't like having a big
output cap for the LDO without having something similar on the input
side too.

So my "let's not use electrolytics" model for the digital base board
needs to go.  But it's still true that this thing depends on the
TPS73801 for reverse voltage protection, and while maybe it would be
saner to just do a diode and a cheap regular LDO, that's not what this
is.

I'm just brute-forcing it, by getting a non-polar electrolytic 100uF
cap.  Is it bigger and more expensive than a regular one? Yes.  Do I
look like I care? No.  It's not like I'm cost-optimizing this for mass
production.  I'm doing single boards.

Is this going to matter? I don't know.  I don't even have any boards
with the PT2399 back yet.  But gosh-darn-it-to-heck, it's the right
thing to do.

Small nit: those caps will have to be put in lying down horizontally,
because they are too big and manly to fit under the effects board.  And
kicad doesn't have a footprint variant for that.  And I'm much too lazy
to make a custom footprint just for that.

So just leave enough space on the board, and try to remember when the
time comes to populate things (or when modifying the board design in the
future).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>