More tweaking of the 5V (aka "digital") base board
I'm starting to think this is really what I should move over to, but
let's take all the good parts from the original VGND board.

My test board is much noisier than the latest VGND boards, and while
much of it is probably because I tried to initially avoid the
electrolytic caps, I do think the ferrite bead on the incoming power
ground may have actually mattered.

Because the noise looks somewhat similar to what the older boards looked
like on the oscilloscope, and things did get cleaner.  And the noise on
the new board shows up even in bypass mode, which implicates the ground
plane to me.

So let's bring over the ferrite on the power ground from the other
board, and try to make this board have cleaner power rails.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>