Quick dirty fix for tremolo board
Bah. This didn't show up in simulation, because I'm stupid.

My "clever" trick to make the LFO output stay negative by using GND as
the upper voltage rail was absolutely moronic.  Why? Because in the VGND
model, GND is not really a power rail at all, just a reference voltage.

Sure, it's a pretty strong reference with the TLE2426 being able to put
some effort into it, but not *that* strong.  It's like 20mA max (I'm not
sure where I got the 80mA number that I have in the VGND schematic, but
it's wrong).

Anyway, there's a much better way to do this - tie the top of the LFO up
to the actual positive power rail, but with a 1k resistor.  That creates
a voltage divider with the resistors to the capacitors, and now we're
not stressing that virtual ground.

And the simulation setup doesn't know that GND is current-limited, so it
didn't show up there, and only when I got a test board back did I go
"why is GNDREF wiggling around like mad? Oh..."

This is a very quick hack - I should think about this more, but even if
it is the right solution and it works I should lay the board out anew.
Instead I just did a hacky minimal re-routing for the nets that changed.
Ugly, but I won't bother doin ga better job unless this actually works
and fixes things.

The good news is that apart from the 10kHz GNDREF oscillation the new
Tremolo board seemed to work.  Of course, that noise could be hiding
other problems, and maybe there's something else wrong.  So let's not
celebrate too early.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>