Well, the VIC-20 board is back, and it's as glorious as expected
.. which is obviously not very glorious, but it does work pretty much
exactly as designed.

Here are two sketches to show off the board.  A plain "read ADC, write
DAC, be slow", and a slightly fancier variant that does the same thing
but buffering it for 128ms.

Not useful, but that was never the point.

I also did some silly test patterns, and the DAC performance was even
worse than I expected.  Or my quick tests were broken, which is entirely
possible too.

I obviously used the Arduino IDE with megaTinyCore, and set the board
and chip to ATtiny814.  This board is undebuggable since it only has an
input pin and the DAC output, but there is also less to debug that way.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>