Start converting effects boards to the 5V base board: tremolo
The original effects boards were designed to not actually care too much
about the details of the voltage rail, but that was never true of the
tremolo board - because the transistor V_{GS(off)} voltage matters.

Yet the tremolo board was made to _look_ like it didn't care about the
details of the rails, having the non-specific +V and -V rails, and
calling the signal ground reference "GND".

That signal ground vs power ground thing obviously then caused the
additional issue with the schematic using the signal voltage reference
as a power rail for the LFO, fixed in commit 63201d9 ("Quick dirty fix
for tremolo board").

The digital effects boards already made their voltage rails explicit,
and match the base board they require.  Let's just start doing that with
the other effects boards too, starting with the tremolo that cares the
most.

Since I have to start using the 5V rail-to-rail op amps anyway, and I
don't want to have lots of component variations for different boards, I
think I'm going to convert them all to the new world order.

And while 9V may be traditional, a lot of modern components seem to all
be geared to 5V or less.  Of course, there are also op amps like the
OPAx991 family that have very wide voltage rails indeed, and are RtR
from 2.7V to 40V rails.  But with other components having much more
limited voltage ranges, I don't think it's worth designing different
base boards.

For the tremolo board, the only real change for 5V involves removing the
voltage divider resistor that was part of driving the JFET.  With the 9V
voltage rails, that just intentionally limited the gate voltage swing.

Now, with the 5V rails, it ended up limiting the gate voltage too much,
and the JFET might not turn on at all depending on exactly where
V_{GS(off)} ends up being.

And since we're now being honest about what is real ground, just make
the shield be GND and use a full ground pour.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>