Update enclosure drill hole details
I've made most of the holes a bit bigger, because while everything fit,
it was just needlessly tight, so inserting and removing boards for
testing was fiddly.

Also, the holes for the pots in particular are now larger, to give me
the option to use the pots with clear plastic shafts that you can put a
SMD LED under and light from below (eg Tayda SKU A-7795).  That might
make sense particularly for the tremolo board.  Those pots want a hole
diameter of 7.5mm.

The regular pots have a ring and nut, so the hole size is not very
critical.

Also, the LED has moved in a bit, and the hole for it has been expanded.
The old location _worked_, but the bezel nut on the inside was really
too close to the enclosure wall.  And the hole was made bigger to also
allow using the plain plastic LED holder.

Finally, the audio jack holes were not only expanded a bit, they also
move up.  The old hole position didn't allow the base board to move far
enough into the enclosure.

As with the pots, the audio jacks have a ring and nut, so any small
variation in size and location isn't actually visible.  And I had a
board that failed because tightening the stomp switch nut pulled the PCB
in enough to then put strees on the audio jack solder joints.

All of these are tiny changes, and the old enclosures actually worked
quite well.  But this should make things more flexible and make
inserting and removing the boards easier.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>