Turn the delay pedal into an echo
For testing, the delay pedal did a pure delayed output.  And it tests
fine, but I think what is more interesting to use is an actual echo
pedal where the delay gets fed back into the input.

Maybe it could have a mixed mode and do either? Or both? Anyway, this
just does the feedback thing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>