Update the VGND board for things from the digital board
This brings the VGND board up to where the digital board was, and also
adds a common mode choke for more power cleaning testing (I know I said
I should do it for the USB-C power board, and I will, but I did it here
too just because I was updating things).

I think I'll try to actively hunt for a couple of cheap and hopefully
bad wall wart guitar power supplies.  Considering how many people
complain about bad guitar power supplies, it should hopefully be easy.

Will the choke matter? I have no idea, but hey, that's kind of the
point, isn't it?

This also adds in the simulation so that you can see the signal
rebiasing happening.  As usual ltspice randomly gives errors that come
and go depending on "use initial conditions" and timing, but when it
works it works.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>