Update all remaining effects boards to use global ground and GNDREF
Let's just make it the rule that we only call the actual external ground
GND, and don't play with local grounds.  That makes things like the
potentiometer shield pins so much simpler to deal with, and means that
ground pours are actual ground pours.

The GNDREF uses are fairly few in the end, and it's much better to make
them explicit.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>