VGND: Add capacitor to GNDREF for stability
The base boards should use GNDREF purely as a voltage reference, but
let's face it, they haven't always done that.  Hopefully no longer
calling it GND will keep it from happening, but this adds a small
capacitor to it in the hope that it will help keep it stable.

The TLE2426 datasheet has a graph of the stability range on pg 17 on
load current vs load capacitance, and that says that it should be stable
with a load capacitance in the 1uF-100uF range for all valid output
currents (+-20mA).

So adding a 10uF capacitor on the base board should be the safe thing to
do.  There's nothing we can do if the effect board just pulls or sinks
more current than the 20mA that the TLE2426 supports, but we can at
least stay off the documented unstable region.

[ From a quick look through the effect boards, it looks like we should
  be good. GNDREF is typically used with the inverting input of an
  op-amp. Or alternatively with sufficiently big resistors to the
  signal, ie tens of kOhms or more. Fingers crossed. ]

Also fix the incorrect commentary about 80mA current capability on the
schematic sheet.  I wonder where I picked that incorrect value from.  I
must have been looking at the wrong part.  Or on some bad meds.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>