Add a Würth reverse-mounted LED footprint
Since I often end up having the SMT components on the bottom layer, with
the connectors and controls on top, the status LEDs end up being either
fairly hard to see, or have to use through-hole components.

Unless - and hear me out - I just start using reverse-mount LEDs where
the LED shines through a hole in the PCB.

The Würth ones look good.  They are kind of normal 1206 size, and you
can mount them that way.  But really, the point is that you can just do
a reverse mount with a 2.1mm hole in the middle too.

They come in red/green/yellow, all with the same 2V forward voltage
drop, but the red one is (as is tradition) more efficient, so probably
do a resistor double the size for that.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>