Add OPA1664 quad op amp symbol
I might trial this for the rectifier board, just to see if doing a
TSSOP-14 quad opamp ends up being convenient.

For the base board, I think the OPA1656 is the right choice, not only
because it only wants a dual op amp anyway, but because it's a high
impedance JFET input with low current noise.

But for the effects boards you might want a quad opamp and the signal
impedance isn't the same kind of issue.

The point? Just more experimentation.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>