Power board: move common mode choke to front
I don't do the nice 3D models, because I'm lazy and it's too much of a
hassle.  But it also means that it wasn't visually obvious that the
common mode choke ends up being very inconvenient on the back of the
board.

I'm perfectly fine with the usual low profile SMD components there, but
that choke is 3.5mm high, which is about twice as much as the other
components.

It's not that bad on the power board (which I don't have an enclosure
for anyway, although I'm looking at maybe just encasing it in resin with
a mold).  On the VGND effects board it's a non-starter.

Of course, I'm not convinced the choke actually matters, and the answer
may just be to remove it again, but until I see the effects of it let's
just move it to the front of the board.

It's going to be inconvenient to solder - I much prefer the hotplate -
but particularly on the power board I already have that USB-C connector
that I do with a hot-air station, and they are right next to each other,
so I can probably just do them together.

I'll have to think about the effects board more.  I don't actually have
room on the top of the board because I want to keep the electrolytics up
there, and keep the area in the middle clear for if I ever re-introduce
the joiner/splitter model with a third audio jack.

I think the inter-board connector there in the middle was a mistake.
It's in the way not only for this choke, but for any single-pot design
if I ever do one.

So the solution may be that I'll just have to move the connector around,
but that obviously means that I lose backwards compatibility with
existing effects boards.  Not really a big deal - unlike the enclosure
changes, it's easy enough to just do new boards - but if I make that
kind of change I want to make sure I think it through and get it right.

So let's wait to see if the choke even makes a difference.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>