Tremolo: move LED to under depth pot
My 3-pot enclosures no longer have the hole for the led to the right of
the stomp switch, so while the LED was still useful for testing, it
wasn't visible inside the enclosure.

It won't be visible with the normal depth pots either, but this way I
can at least experiment with the clear shaft pots and make the tremolo
speed visible that way.

Also update some of the component values based on more testing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>