Simplify base board GND noise isolation
I finally ended up building multiple test boards to actually check the
experimental parts of the ground isolation.

It turns out none of it helps reduce noise, and the actual ground loop
isolation resistors actively hurt.

It also turns out that the biggest source of noise is actually my
soldering station.  Turning that off when looking at the signals on the
oscilloscope makes quite a noticeable difference ;)

Maybe the ground loop resistors might help in some special situations,
but when looking at a signal generator on the scope they most definitely
hurt.  Having good grounding everywhere is just better.

And I much prefer the "ground everything in a good ground plane" over
the alternatives, particularly since it simplifies things too.  So let's
end that experiment.

Also removing the common mode choke and the ferrite on the ground input
path.  They weren't noticeable in any test, and I think improving the
power supply side is the better option.

Finally - I like the reverse-mounted LEDs, but for the base board they
end up eventaully invisible, and since we need to have the "real" LED
for the stomp switch indicator, they are cute but not really productive.
So they are gone again.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>