Update footprint for the 6.35mm mono jack
This was one of the first footprints I did, and I did it oddly, with the
center point of the footprint in a strange random place.

Also, some of the measurements for the courtyard and silkscreen weren't
quite right, which I think was a result of partly doing it with calipers
and partly looking at the datasheet.  The fact that the origin for the
footprint was odd made it not very obvious.

So center the footprint on the central ground pin, which is what most of
the datasheet measurements are referenced off.

And while those issues were mostly cosmetic, there were a couple of real
issues with the footprint too.  The hole for the tip and switch pad was
annoyingly tight: the datasheet says they are 2.5mm wide, and that was
the exact size of the hole.  Yes, it worked, and certainly kept the
component in place, but it was always a slight annoyance.

A bigger annoyance was that because the 1590A build is so tight
vertically, I always ended up snipping off the plastic standoffs / feet
on the jack.  They make the jack be 12mm high, and cutting them off
saves about 0.75mm of vertical space.

But I did it mostly because I didn't want to bother with the footprint
definition, and partly because the alternative to cutting off the feet
was to have mechanical holes in the PCB, and I felt that board space was
too valuable.

Now, it turns out that board space isn't all that valuable on the base
board - there just aren't that many components there - and since I
wanted to fix up the other mostly cosmetic issues, I felt that not
having to use flush cutters on the jacks when assembling it is the
better option.

This all means that the boards that have these footprints need to be
updated for the re-positioning and the mechanical holes that mean that
you can't place components and traces on either side.  But it's really
just the regular base board and the trivial debug board that need some
TLC.

While doing that, I also moved the connectors out even further.  This is
a further tweak of commit 49b93c2 ("Move mono connectors out by 0.8mm
each"), because it turns out that the threads at the end of those
connectors really aren't very good.

And no, those audio connectors still don't stricly need the nut and
ring, but with the undercut of the hole in the enclosure I really prefer
having them, even if it means that the whole connector sticks out more
that way.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>