Add 'joiner' board for blending two incoming lines
This was part of the original designs in the legacy branch, but never
got added in the Gen2 boards because it needs a different enclosure
setup.  But it's still something I want to have for experimentation with
different combinations, so let's add it back.

In particular, I think I'll make an "exploded" version of the rectifier
board that does the individual parts of it, because Dirk actually liked
the sound of the original Octaver board for his bass guitar.

But that old circuit is very noisy, and doesn't simulate at all.  It
also had had a very different tone control (plain Baxandall) and joined
the dry and wet signals before tone control.  So it's hard to even guess
about what part of it made him like it without splitting these different
parts up.

Which obviously means I need a splitter board too - and the individual
split out tone control and plain rectifier boards.  That's up next.

[ Side note: this is a single-pot board, which means that with the
  current inter-board connector layout we need some funky positioning
  for the pot.

  As mentioned elsewhere, I really need to re-think the whole
  inter-board thing, but at least for now we work around it with
  strangely rotated parts ]

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>