Add Baxandall tone effect board
This is one more board in the "separate out the circuits of the
rectifier boards" series, so that I can find out what part of the
original Octaver circuit that makes Dirk like it for the bass guitar
setup.

And honestly, I should have done it this way originally.

This time I also made sure to simulate it properly, and fix the
oscillations that the original board had that made me dislike it so much
(famous last words - let's see how well simulation matches reality, but
yes, the original rectifier board did oscillate horribly).

I think this is what I should do for the boost effect too.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>