Add a new pure rectifier effect board
This is a slightly different rectification circuit that enters at the
inverting input of the first op amp - which also allows for a gain of
two while doing the rectification.

That way the amplitude of the output is the the same as the amplitude of
the input assumiong it's centered around ground.

But perhaps even more importantly, it seems to simulate a lot more
reliably, and I haven't needed to fight ltspice nearly as much.  Maybe
pure luck, but I don't think I've seen a single 'timestep too small'
error so far.

Knock wood.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>