Add volume control to 'splitter' board
Also stop being silly: this board needs no active circuitry.  The
incoming signal already comes from an op amp buffer on the base board,
there's no need to add a second one.

So remove the pointless op-amp, and replace the 100k resistor to ground
that is part of the signal rebiasing with a potentiometer for volume
control.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>