Add 6.35mm stereo jack footprint
I have a few of these, because I at some point thought that I'd do the
splitter thing by just using a stereo jack instead of two mono jacks.

Maybe I'll use these some day, but I decided against it for now because
it's just such a pain to mix mono and stereo jacks.  If you put a
regular patch cable (mono) into a TRS stereo jack, it will short the
ring (ie typically right channel) to ground through the sleeve.

So the obvious and convenient thing to do with a mono board - to just
say "connect tip and ring together" - doesn't work, because then both
will just be shorted to ground if you use a mono cable.

Could I make the base board be a proper stereo board? At the cost of
just one mostly unused op-amp, if you put a mono cable in the right
channel just gets shorted to ground but the left channel works fine.

And since you'd need an extra op-amp, you'd use a quad op-amp chip, and
at that point you might as well do stereo input too, even if the right
channel would be just silent when you have a mono input.

Is there some clever way to just recognize the short-to-ground situation
and automatically turn mono input into "pseudo-stereo" with L/R being
tied to the same signal?

Of course, to make that stereo situation useful for anything but the
splitter/joiner cases, I'd have to pass the signals to the effects
boards.  Which won't happen until I figure out what I really want the
board connection situation to be.  Or have a board that would ever use
that.

The big advantage of trying to be clever about this would be to avoid
the enclosure situation with the extra input/output jacks.

Anyway.  This does none of that.  This just adds the footprint that is
currently not used or useful.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>