Baxandall: remove unnecessary output resistors
These came from the TI reference design, and I kept them around until I
actually put the board together, at which point I went "I like the
tinkering, but not pointlessly".

So remove them.

I'm not sure why that 47kΩ resistor to ground was there.  As some kind
of preemptive DC biasing for the next stage with a capacitive input? Not
relevant for this design.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>