VGND: move THT capacitor up for more room
So the latest round of test builds found a small issue with the base
board: the 100uF THT capacitor needs to lie down on the board, because
the ones I have are too tall at 13mm to stand up straight (11.5mm
between boards).  That was all according to plan.

What was also according to plan was leaving enough space for the extra
audio input or output jack for the joiner/splitter boards.  And that was
a "kind of".  It all worked, but to work well you had to really bend the
capacitor legs so that it moved up on the board before it then lay down
flat on it.

So let's just move the through-holes up a couple of mm so that the
capacitor doesn't need to do some kind of downward dog yoga stretching
thing.

This comes back to the issue that I really should have made a footprint
for the capacitor lying down to make it much more visually obvious,
instead of noticing that things are a bit awkward on the board when
putting it together.

I did know it was a bit tight, and it wasn't really a big problem in
practice.  But it's easy to fix.  Let's be proactive about it so that
next time I build boards I won't have the poor capacitors do
contortions.

The good news? The current boards really do look better noise-wise.  I
noticed it when comparing with one of the previous-gen ones side by side
on the scope while doing the splitter board.  So this really is a fairly
small cosmetic fix.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>