Add initial attempt at OpenSCAD model 1590A enclosure
Note the "initial".  And "attempt", for that matter.

I haven't actually tried printing this yet, and it's my first time ever
using OpenSCAD, but I do think I prefer this kind of model over having
to draw things in 3D.

It looks somewhat reasonable, but who knows how usable it is.

There's no lid, no screw holes, nothing like that.  This is for testing
and for test-fitting.  So it tries to approximate the right inside and
outside dimensions and have the cutouts for knobs and stuff in the right
places.

Screw holes and a lid can be added later if this actually ends up
reasonable.  Although I suspect a plastic print isn't sturdy enough, but
maybe that could be fixed by filling it with epoxy?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>