Make the USB-C power enclosure more robust
The lip holding the lid and body together was a bit flimsy, so replace
it with sturdier clips in the corners and add proper screw holes for the
posts.

This seems very solid and looks pretty good.  The lack of visible status
LEDs is a bit sad, but let's ignore it for now.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>