OpenSCAD 1590A enclosure: make it easier to configure
This just makes the location of the pots and pans be variables, so that
it's easy to just say "I want a three-knob splitter enclosure with a
switch cutout".

Also changed the lip definition to be a corner clip - that worked well
for the USB-C power enclosure, so do that here too.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>